+++ packages/@webex/plugin-meetings/src/meeting/index.ts
@@ -124,7 +124,6 @@ import RecordingController from '../recording-controller';
 import ControlsOptionsManager from '../controls-options-manager';
 import PermissionError from '../common/errors/permission';
 import {LocusMediaRequest} from './locusMediaRequest';
-import {AnnotationInfo} from '../annotation/annotation.types';
 
 const {isBrowser} = BrowserDetection();
 
@@ -150,7 +149,6 @@ export type LocalTracks = {
     audio?: LocalSystemAudioTrack;
     video?: LocalDisplayTrack;
   };
-  annotationInfo?: AnnotationInfo;
 };
 
 export type AddMediaOptions = {
@@ -551,7 +549,6 @@ export default class Meeting extends StatelessWebexPlugin {
   roles: any[];
   environment: string;
   namespace = MEETINGS;
-  annotationInfo: AnnotationInfo;
   allowMediaInLobby: boolean;
 
   /**
@@ -3032,20 +3029,14 @@ export default class Meeting extends StatelessWebexPlugin {
    * @returns {boolean}
    */
   private arePolicyRestrictionsSupported() {
-    // Locus calls do not return the correct display hints
-    if (this.isLocusCall()) {
+    // If we don't have policies we can't support policies
+    if (!this.selfUserPolicies) {
       return false;
     }
 
     // 1-2-1 calls and SIP dialling will have no meeting info
     // so cannot support policy information
     if (isEmpty(this.meetingInfo)) {
-      return false;
-    }
-
-    // Old locus info api does not return policy information
-    // @ts-ignore
-    if (!this.config.experimental.enableUnifiedMeetings) {
       return false;
     }
 
@@ -5430,6 +5421,10 @@ export default class Meeting extends StatelessWebexPlugin {
           url: this.deviceUrl,
           // @ts-ignore
           deviceType: this.config.deviceType,
+          // @ts-ignore
+          countryCode: this.webex.meetings.geoHintInfo?.countryCode,
+          // @ts-ignore
+          regionCode: this.webex.meetings.geoHintInfo?.regionCode,
         },
         preferTranscoding: !this.isMultistream,
       },
@@ -5441,9 +5436,6 @@ export default class Meeting extends StatelessWebexPlugin {
 
     this.audio = createMuteState(AUDIO, this, audioEnabled);
     this.video = createMuteState(VIDEO, this, videoEnabled);
-
-    this.annotationInfo = localTracks?.annotationInfo;
-
     const promises = [];
 
     // setup all the references to local tracks in this.mediaProperties before creating media connection
@@ -5923,20 +5915,25 @@ export default class Meeting extends StatelessWebexPlugin {
     /// @ts-ignore
     this.webex.internal.newMetrics.submitInternalEvent({name: 'internal.reset.join.latencies'});
 
-    // @ts-ignore
-    this.webex.internal.newMetrics.submitClientEvent({
-      name: 'client.call.leave',
-      payload: {
-        trigger: 'user-interaction',
-        canProceed: false,
-        leaveReason,
-      },
-      options: {meetingId: this.id},
-    });
+    const submitLeaveMetric = (payload = {}) =>
+      // @ts-ignore
+      this.webex.internal.newMetrics.submitClientEvent({
+        name: 'client.call.leave',
+        payload: {
+          trigger: 'user-interaction',
+          canProceed: false,
+          leaveReason,
+          ...payload,
+        },
+        options: {meetingId: this.id},
+      });
     LoggerProxy.logger.log('Meeting:index#leave --> Leaving a meeting');
 
     return MeetingUtil.leaveMeeting(this, options)
       .then((leave) => {
+        // CA team recommends submitting this *after* locus /leave
+        submitLeaveMetric();
+
         this.meetingFiniteStateMachine.leave();
         this.clearMeetingData();
 
@@ -5972,6 +5969,20 @@ export default class Meeting extends StatelessWebexPlugin {
         return leave;
       })
       .catch((error) => {
+        // CA team recommends submitting this *after* locus /leave
+        submitLeaveMetric({
+          errors: [
+            {
+              fatal: false,
+              errorDescription: error.message,
+              category: 'signaling',
+              errorCode: 1000,
+              name: 'client.leave',
+              shownToUser: false,
+            },
+          ],
+        });
+
         this.meetingFiniteStateMachine.fail(error);
         LoggerProxy.logger.error('Meeting:index#leave --> Failed to leave ', error);
         // upload logs on leave irrespective of meeting delete
@@ -6155,7 +6166,6 @@ export default class Meeting extends StatelessWebexPlugin {
             deviceUrl: this.deviceUrl,
             uri: content.url,
             resourceUrl: this.resourceUrl,
-            annotationInfo: this.annotationInfo,
           })
           .then(() => {
             this.screenShareFloorState = ScreenShareFloorStatus.GRANTED;
@@ -7033,9 +7043,6 @@ export default class Meeting extends StatelessWebexPlugin {
    */
   async publishTracks(tracks: LocalTracks): Promise<void> {
     this.checkMediaConnection();
-
-    this.annotationInfo = tracks.annotationInfo;
-
     if (
       !tracks.microphone &&
       !tracks.camera &&