+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0-beta.211](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.210...v3.0.0-beta.211) (2023-09-11)
+
+
+### Features
+
+* **call-leave:** reorder metric send ([#3007](https://github.com/webex/webex-js-sdk/issues/3007)) ([7593198](https://github.com/webex/webex-js-sdk/commit/7593198e243076b7ab433d9f65116c872f33b81b))
+
+## [3.0.0-beta.210](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.209...v3.0.0-beta.210) (2023-09-08)
+
+
+### Features
+
+* **plugin-meetings:** fix canDoVideo ([#3063](https://github.com/webex/webex-js-sdk/issues/3063)) ([f65d73b](https://github.com/webex/webex-js-sdk/commit/f65d73b18cf123bab117715d0967e182065f3c9f))
+
+## [3.0.0-beta.209](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.208...v3.0.0-beta.209) (2023-09-08)
+
+
+### Bug Fixes
+
+* geohint info missing in Locus join requests ([#3061](https://github.com/webex/webex-js-sdk/issues/3061)) ([400286b](https://github.com/webex/webex-js-sdk/commit/400286bc576882999396cc48ac266cea13f98f2d))
+
+## [3.0.0-beta.208](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.207...v3.0.0-beta.208) (2023-09-07)
+
+
+### Bug Fixes
+
+* **Live-Annotation:** remove the presenter side annotation info ([#3048](https://github.com/webex/webex-js-sdk/issues/3048)) ([3e5e800](https://github.com/webex/webex-js-sdk/commit/3e5e80061be835c55e18216b01079c8b4ecd860c))
+
+## [3.0.0-beta.207](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.206...v3.0.0-beta.207) (2023-09-06)
+
+
+### Features
+
+* **ca:** fix prod meta ([#3060](https://github.com/webex/webex-js-sdk/issues/3060)) ([b359f26](https://github.com/webex/webex-js-sdk/commit/b359f26d157f9b40c25ea011aaef80675bda4469))
+
 ## [3.0.0-beta.206](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.205...v3.0.0-beta.206) (2023-09-06)